Refactor polyline measure to be its own component
